requirements-dev.txt
coverage==7.8.2
Deprecated==1.2.18
iniconfig==2.1.0
lxml==5.4.0
packaging==25.0
piexif==1.1.3
pikepdf==9.8.1
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.1
pypdf==5.6.0
pytest==8.4.0
pytest-cov==6.1.1
wrapt==1.17.2
